# openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Nexus REST API",
    "description": "Explore resources and get their data.",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/api/v1/catalogs/{catalogId}": {
      "get": {
        "tags": [
          "Catalogs"
        ],
        "summary": "Gets the specified catalog.",
        "operationId": "Catalogs_Get",
        "parameters": [
          {
            "name": "catalogId",
            "in": "path",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceCatalog"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/authenticate": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Authenticates the user.",
        "operationId": "Users_Authenticate",
        "parameters": [
          {
            "name": "scheme",
            "in": "query",
            "required": true,
            "description": "The authentication scheme to challenge.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "returnUrl",
            "in": "query",
            "required": true,
            "description": "The URL to return after successful authentication.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/signout": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Logs out the user.",
        "operationId": "Users_SignOut",
        "parameters": [
          {
            "name": "returnUrl",
            "in": "query",
            "required": true,
            "description": "The URL to return after logout.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/users/tokens/refresh": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Refreshes the JWT token.",
        "operationId": "Users_RefreshToken",
        "requestBody": {
          "x-name": "request",
          "description": "The refresh token request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "A new pair of JWT and refresh token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/tokens/revoke": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Revokes a refresh token.",
        "operationId": "Users_RevokeToken",
        "requestBody": {
          "x-name": "request",
          "description": "The revoke token request.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeTokenRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/tokens/generate": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Generates a refresh token.",
        "operationId": "Users_GenerateRefreshToken",
        "parameters": [
          {
            "name": "description",
            "in": "query",
            "required": true,
            "description": "The refresh token description.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "userId",
            "in": "query",
            "description": "The optional user identifier. If not specified, the current user will be used.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/accept-license": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Accepts the license of the specified catalog.",
        "operationId": "Users_AcceptLicense",
        "parameters": [
          {
            "name": "catalogId",
            "in": "query",
            "required": true,
            "description": "The catalog identifier.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/tokens/{tokenId}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a refresh token.",
        "operationId": "Users_DeleteRefreshToken",
        "parameters": [
          {
            "name": "tokenId",
            "in": "path",
            "required": true,
            "description": "The identifier of the refresh token.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{userId}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a user.",
        "operationId": "Users_DeleteUser",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The identifier of the user.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}/claims": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Gets all claims.",
        "operationId": "Users_GetClaims",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The identifier of the user.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/NexusClaim"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Creates a claim.",
        "operationId": "Users_CreateClaim",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The identifier of the user.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "claim",
          "description": "The claim to create.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NexusClaim"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "guid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/claims/{claimId}": {
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Deletes a claim.",
        "operationId": "Users_DeleteClaim",
        "parameters": [
          {
            "name": "claimId",
            "in": "path",
            "required": true,
            "description": "The identifier of the claim.",
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/{userId}/tokens": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Gets all refresh tokens.",
        "operationId": "Users_GetRefreshTokens",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "description": "The identifier of the user.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "$ref": "#/components/schemas/RefreshToken"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResourceCatalog": {
        "type": "object",
        "description": "A catalog is a top level element and holds a list of resources.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets the identifier."
          },
          "properties": {
            "type": "object",
            "description": "Gets the properties.",
            "nullable": true,
            "additionalProperties": {}
          },
          "resources": {
            "type": "array",
            "description": "Gets the list of representations.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Resource"
            }
          }
        }
      },
      "Resource": {
        "type": "object",
        "description": "A resource is part of a resource catalog and holds a list of representations.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "description": "Gets the identifier."
          },
          "properties": {
            "type": "object",
            "description": "Gets the properties.",
            "nullable": true,
            "additionalProperties": {}
          },
          "representations": {
            "type": "array",
            "description": "Gets the list of representations.",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/Representation"
            }
          }
        }
      },
      "Representation": {
        "type": "object",
        "description": "A representation is part of a resource.",
        "additionalProperties": false,
        "properties": {
          "dataType": {
            "description": "The data type.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/NexusDataType"
              }
            ]
          },
          "samplePeriod": {
            "type": "string",
            "description": "The sample period.",
            "format": "duration"
          },
          "parameters": {
            "type": "object",
            "description": "The optional list of parameters.",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
      "NexusDataType": {
        "type": "string",
        "description": "Specifies the Nexus data type.",
        "x-enumNames": [
          "UINT8",
          "UINT16",
          "UINT32",
          "UINT64",
          "INT8",
          "INT16",
          "INT32",
          "INT64",
          "FLOAT32",
          "FLOAT64"
        ],
        "enum": [
          "UINT8",
          "UINT16",
          "UINT32",
          "UINT64",
          "INT8",
          "INT16",
          "INT32",
          "INT64",
          "FLOAT32",
          "FLOAT64"
        ]
      },
...
      "TokenPair": {
        "type": "object",
        "description": "A token pair.",
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "The JWT token."
          },
          "refreshToken": {
            "type": "string",
            "description": "The refresh token."
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "description": "A refresh token request.",
        "additionalProperties": false,
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token."
          }
        }
      },
      "RevokeTokenRequest": {
        "type": "object",
        "description": "A revoke token request.",
        "additionalProperties": false,
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token."
          }
        }
      },
...
      "RefreshToken": {
        "type": "object",
        "description": "A refresh token.",
        "additionalProperties": false,
        "properties": {
          "expires": {
            "type": "string",
            "description": "The date/time when the token expires.",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "description": "The token description."
          }
        }
      },
      "NexusClaim": {
        "type": "object",
        "description": "Represents a claim.",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "description": "The claim type."
          },
          "value": {
            "type": "string",
            "description": "The claim value."
          }
        }
      }
    }
  }
}